fix(chat): harden side question coordination
